master_node/requirement.txt




PyPDF2
flask
llama-index
llama-index-embeddings-langchain
langchain
langchain_community
jupyter
notebook
pandas
numpy
tqdm

web3
flask_cors
flask-csp
Flask-Talisman
gunicorn
PyJWT

sentence-transformers

python-dotenv